Major refactor, created `Game` class, separated header files into `include` directory and tidied up `CMakeLists.txt`
