+++ .clang-format
@@ -1,0 +1,3 @@
+# Generated by CLion for LLVM
+BasedOnStyle: LLVM
+IndentWidth: 4
