+++ src/game/Player.cpp
@@ -1,8 +1,9 @@
 //
-// Created by Jacopo Uggeri on 15/08/2025.
+// Created by Jacopo Uggeri on 20/08/2025.
 //
 
-#include "GameState.h"
+#include "game/Player.h"
+#include "game/physics/Vec2.h"
 
 void Player::set_velocity(phys::Vec2f v) {
     vel = v;
@@ -11,8 +12,4 @@ void Player::set_velocity(phys::Vec2f v) {
 void Player::move(double deltaSeconds) {
     const auto dt = static_cast<float>(deltaSeconds);
     pos +=  vel * SPEED * dt;
-}
-
-void GameState::step(double deltaSeconds) {
-    player.move(deltaSeconds);
 }
